refactor: unify frame loop into process_frame_any
Replace the 3-arm match in run_immediate (GpuResident, D3d11Resident,
catch-all) with a single process_frame_any method that guarantees
detect -> pose -> render -> replay -> telemetry fires for every
StereoFrame variant.

run_immediate is now 39 lines: decode + event + process_frame_any +
progress. step_gpu_with_bufs inlined into the GpuResident arm and
deleted.

Adds FrameLoopContext to carry pre-cloned GPU buffer info across the
borrow checker boundary (150 bytes, cloned once before the loop).

Fixes pre-existing telemetry bug: detection timing flag was read
after submit_render_output incremented frame_count (off by one with
detection_interval > 1). Now captured before render.

All paths get full telemetry (decode + detection + stitch + total).
Previously GpuResident and D3d11Resident only recorded decode + total.

135 tests pass, clippy clean (including --features profiling).
